Add support for profiling theimage with Java Flight Recorder
