+++ demo/locale-plugin/repo/pom.xml
