+++ demo/locale-plugin/repo/src/main/resources/hudson/plugins/locale/PluginImpl/config.properties
