+++ demo/locale-plugin/repo/src/main/resources/hudson/plugins/locale/PluginImpl/config_zh_TW.properties
