+++ demo/locale-plugin/repo/src/main/resources/index.jelly
@@ -1,0 +1,4 @@
+<?jelly escape-by-default='true'?>
+<div>
+  This plugin lets you control the language of Jenkins.
+</div>
