+++ demo/locale-plugin/repo/src/test/java/hudson/plugins/locale/MigrationTest.java
