+++ demo/locale-plugin/repo/src/test/resources/hudson/plugins/locale/MigrationTest/dataMigration_13/locale.xml
