--- demo/locale-plugin/src/main/resources/hudson/plugins/locale/PluginImpl/config.jelly
@@ -1,12 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <f:section title="${%Locale}">
-        <f:entry title="${%Default Language}" help="/plugin/locale/help/default-language.html">
-            <f:textbox name="systemLocale" value="${it.systemLocale}" />
-        </f:entry>
-        <f:nested>
-            <f:checkbox name="ignoreAcceptLanguage" checked="${it.ignoreAcceptLanguage}" />
-            <label class="attach-previous">${%description}</label>
-        </f:nested>
-    </f:section>
-</j:jelly>
