--- demo/locale-plugin/src/main/resources/index.jelly
@@ -1,4 +1,0 @@
-<?jelly escape-by-default='true'?>
-<div>
-  This plugin lets you control the language of Jenkins.
-</div>
